Unbreak the levels plugin's “luma” mode.
When used in the default “luma” mode, levels plugin was effectively trying to
approximate the value-to-value map, including pow, subtractions, etc., by a
single multiplication (except in the case of all-zero input, in which case
it would simply correctly look up in the map). The calculated luma value would
then be used to decide which multiplication factor to use.

This is obviously not going to work; the typical result would be that
when using black output level > 0, very dark but non-zero pixels (e.g. #000100)
would be amplified a lot and create colored artifacts in such areas.
Simply let the “luma” mode apply the curve equally to all color channels,
which is probably what the user expected. The histogram still shows the
actual calculate luma, though.

There's a lot of performance work to do on this plugin, e.g. move the
switch out of the inner loop, but at least it should give slightly more
sensible output now.